inventory-management-system/pyproject.toml
[project]
name = "inventory-management-system"
version = "0.1.0"
description = "A Streamlit-based Inventory Management System"
requires-python = ">=3.8"
dependencies = [
    "streamlit",
    "pandas",
    "sqlalchemy",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"